requirements.txt
moviepy
SpeechRecognition
coqui-tts
googletrans==4.0.0-rc1
gTTS
streamlit
groq
Pillow
